Refactor; pull Backup interfaces out into their own source files
With introduction of FileTree interface, that file was getting too big.
