+++ file-backup-s3/src/test/java/com/willmolloy/backup/s3/S3FileTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 
 import com.github.javafaker.Faker;
-import com.willmolloy.backup.FileTree;
+import com.willmolloy.backup.File;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -67,7 +67,7 @@ class S3FileTest {
     // Given
     S3File thisFile = spy(new S3File(S3Object.builder().size(thisSize).build()));
 
-    FileTree.File otherFile = mock(FileTree.File.class);
+    File otherFile = mock(File.class);
     when(otherFile.size()).thenReturn(otherSize);
 
     // Then
